Fixed wrong typedef for EE_UINT64 with 16 bits architectures in /pkg/cpu/common/ee_types.h
git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@2609 88d5ba13-baa4-4915-ae7a-41304cef7dd7

+++ pkg/cpu/common/inc/ee_types.h
@@ -66,7 +66,7 @@ typedef long int EE_INT32;
 typedef unsigned long int EE_UINT32;
 #ifdef ULLONG_MAX
 #if ULLONG_MAX == 0xffffffffffffffffUL
-typedef unsigned long int EE_UINT64;
+typedef unsigned long long int EE_UINT64;
 #endif
 #endif
 #elif (INT_MAX == 2147483647) && (SHRT_MAX == 32767) /* 32-bit architecture */
